# api/openapi.yaml
...
components:
  schemas:
    ErrorResponse:
      properties:
        error:
          type: string
      type: object
    MainUsage:
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
      type: object
    ModelUsage:
      properties:
        calls:
          type: integer
        cost_usd:
          type: number
        input_tokens:
          type: integer
        model:
          type: string
        output_tokens:
          type: integer
      type: object
...
    SessionItem:
      properties:
        auth_subject:
          description: |-
            AuthSubject is the gateway-stamped JWT subject (WorkOS user id)
            captured at ingest; empty for rows captured before the edge began
            stamping it.
          type: string
        cwd:
          type: string
        display_name:
          description: |-
            DisplayName is the user's Console rename (sessions.display_name),
            empty unless a user set one. Written only by PATCH /v1/sessions/:id,
            never by ingest, so it survives a live session. It is the top of the
            DisplayTitle resolution; exposed raw so the edit affordance can seed
            its input from the user's own title (not the resolved fallback).
          type: string
        display_title:
          description: |-
            DisplayTitle is the server-resolved label clients should render:
            DisplayName -> rollup.title (generated) -> preview -> Name -> id
            slice. Resolving once on the server keeps every client (Console,
            paper CLI) from re-deriving — and diverging on — the precedence
            (PCC-970). Never empty: it falls back to a short harness id slice, then
            the session id (the primary key, always set for a stored row).
          type: string
        ended_at:
          type: string
        harness_id:
          type: string
        harness_metadata:
          additionalProperties: {}
          type: object
        harness_session_id:
          type: string
        harness_version:
          type: string
        id:
          description: Identity — capture-side facts, ingest-written.
          type: string
        last_seen_at:
          type: string
        live:
          description: |-
            Live is a runtime presence signal, not a projection fact: true when
            the session has no recorded end and was seen within the liveness
            window. Keyed on ended_at + last_seen_at recency (both ingest-fresh),
            never on the derived status: an interactive session folds to a
            terminal status (an end_turn assistant reply reads as "completed")
            after every turn while still open, so status cannot gate liveness.
            Computed at response time so the console renders it directly instead
            of inferring "running" itself.
          type: boolean
        name:
          description: |-
            Name is the harness identity-row label — the harness-supplied session
            name (a plan slug), or the folded title (rollup.title) as a fallback
            when no name was captured. This is capture/deriver provenance, NOT a
            user title: ingest re-sends it every turn. Clients should render
            DisplayTitle, not Name (PCC-970).
          type: string
        parent_session_id:
          type: string
        rollup:
          allOf:
          - $ref: '#/components/schemas/SessionRollup'
          description: Rollup is the deriver-owned projection over the session's spans.
        started_at:
          type: string
      type: object
...
    SessionRollup:
      properties:
        kind_counts:
          additionalProperties:
            type: integer
          description: |-
            KindCounts (spans per call_kind) and Tasks (TaskCreate/TaskUpdate
            folds) are pinned so the rollup shape is uniform across sessions.
          type: object
        model:
          description: |-
            Model is the dominant conversation-spine model; ModelUsage is the
            per-model spend breakdown across every thread (subagent models
            included), cost-ordered so the UI can show "dominant model + share"
            without a cheap-subagent fan-out skewing it.
          type: string
        model_usage:
          items:
            $ref: '#/components/schemas/ModelUsage'
          type: array
        preview:
          type: string
        status:
          type: string
        tasks:
          items:
            $ref: '#/components/schemas/TreeTask'
          type: array
        title:
          description: |-
            Title is the deriver's folded session title (derived_title),
            generated from the conversation. Empty until title generation
            produces one. It never falls back to the identity-row name, so it is
            the stable descriptive title clients prefer for display; the
            identity-row label (harness name or rename) is SessionItem.Name.
          type: string
        turn_count:
          type: integer
        usage:
          $ref: '#/components/schemas/SessionUsage'
      type: object
    SessionTracesResponse:
      properties:
        links:
          items:
            $ref: '#/components/schemas/SpanLinkItem'
          type: array
        schema:
          type: string
        session:
          $ref: '#/components/schemas/SessionItem'
        traces:
          items:
            $ref: '#/components/schemas/TraceDetail'
          type: array
      type: object
    SessionUsage:
      properties:
        cost_usd:
          type: number
        input_tokens:
          type: integer
        output_tokens:
          type: integer
      type: object
    SpanItem:
      properties:
        call_kind:
          description: Deriver-written taxonomy, promoted from the old metadata grab-bag.
          type: string
        duration_ns:
          type: integer
        input:
          description: |-
            Input/Output are content-block arrays (llm.ContentBlock), uniform for
            every kind (tool spans included — no unwrapping). Pinned to [] when
            empty. swaggertype keeps the wire an array-of-object rather than the
            json.RawMessage byte array swag infers.
          items:
            type: object
          type: array
        kind:
          type: string
        model:
          type: string
        name:
          type: string
        output:
          items:
            type: object
          type: array
        parent_span_id:
          type: string
        payload:
          description: |-
            Payload marks a preview-truncated span so the console drills in for
            the full payload; absent in full mode.
          type: string
        raw_turn_id:
          type: integer
        seq:
          description: |-
            Seq is the span's presentation ordinal within its trace; spans
            arrive sorted by it (started_at ties inside one llm call — parallel
            tool batches share an instant).
          type: integer
        span_id:
          type: string
        started_at:
          type: string
        status:
          type: string
        stop_reason:
          type: string
        thread_id:
          type: string
        trace_id:
          type: string
        usage:
          description: |-
            Usage (was `metrics`) is an llm.Usage object on the wire — {}-pinned
            for usage-less spans.
          type: object
        verdict:
          description: |-
            Verdict is the typed security-monitor disposition (null off
            permission-check spans), deriver-written. It is a Verdict object or
            null on the wire; typed as `object` (not the json.RawMessage byte
            array swag would otherwise emit) via swaggertype.
          nullable: true
          type: object
      type: object
    SpanLinkItem:
      properties:
        from_io:
          type: string
        from_span_id:
          type: string
        from_trace_id:
          type: string
        kind:
          type: string
        to_io:
          type: string
        to_span_id:
          type: string
        to_trace_id:
          type: string
      type: object
...
    TraceDetail:
      properties:
        links:
          items:
            $ref: '#/components/schemas/SpanLinkItem'
          type: array
        schema:
          type: string
        spans:
          items:
            $ref: '#/components/schemas/SpanItem'
          type: array
        trace:
          $ref: '#/components/schemas/TraceItem'
      type: object
    TraceItem:
      properties:
        duration_ns:
          type: integer
        ended_at:
          type: string
        main_usage:
          $ref: '#/components/schemas/MainUsage'
        response_preview:
          description: |-
            ResponsePreview is the derive-time fold of the closing
            conversation-spine llm call's text output — the answer line for
            collapsed turn cards, so summary consumers never need spans.
          type: string
        source:
          description: |-
            Source is the capture origin of the turn's rows ("wire" |
            "transcript"), promoted from raw_turns.source. Per-trace, so a
            session can mix live wire capture and transcript backfill. Today
            every trace is "wire": transcripts only reconcile fork/parent edges
            during derivation, they never form a trace on their own. "transcript"
            becomes real when a session is reconstructed purely from a transcript
            file with no proxy capture (an OSS backfill path).
          type: string
        span_count:
          type: integer
        started_at:
          type: string
        status:
          type: string
        synthetic:
          description: |-
            Synthetic is a typed deriver signal ("post-compaction" for a
            compaction continuation, "shadow-opener" for a shadow-only opener),
            promoted out of the old metadata grab-bag. Absent for genuine
            prompt-opened turns.
          type: string
        trace_id:
          type: string
        usage:
          allOf:
          - $ref: '#/components/schemas/TraceUsage'
          description: |-
            Usage is the trace's total token/cost spend over ALL llm spans,
            shadow calls included; MainUsage is the task slice — the main agent
            and its subagents (every call_kind=main span, across threads). The
            difference (Usage − MainUsage) is the harness's shadow spend
            (permission checks, title-gen, web summaries) on the turn.
        user_prompt:
          description: |-
            UserPrompt is served explicitly (not omitempty): a synthetic opener
            has an empty prompt, and dropping the key turns the empty string
            into `undefined` on the wire, which breaks consumers that expect a
            string (e.g. the console's stripHarnessTags). Empty means synthetic.
          type: string
      type: object
...
    TraceUsage:
      properties:
        cache_creation_tokens:
          type: integer
        cache_read_tokens:
          type: integer
        cost_usd:
          type: number
        input_tokens:
          type: integer
        output_tokens:
          type: integer
      type: object
    TreeTask:
      properties:
        description:
          type: string
        id:
          type: string
        status:
          type: string
        subject:
          type: string
        updates:
          type: integer
      type: object
...
info:
  contact: {}
  description: |-
    HTTP API for inspecting, querying, and searching stored Tapes sessions.

    The REST surface exposes health checks, session listing and retrieval, derived session summaries, aggregate stats, semantic search, skill authoring and publishing, operator maintenance endpoints, and a streamable MCP endpoint.

    Not covered here: the ingest write surface, which is a separate server with its own contract (see ingest/openapi.yaml).
  title: Tapes API
  version: "1.0"
openapi: 3.0.3
paths:
...
  /v1/sessions/{id}/traces:
    get:
      description: Returns the session's user-visible turns as traces with nested
        spans (llm calls, tools, subagents, shadow calls, injected context) and dataflow
        links. Cross-trace links (compaction seams) are at the response top level.
      operationId: getSessionTraces
      parameters:
      - description: Session id (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: 'Span payload mode: full (default) or preview (strings truncated;
          fetch the span endpoint for full payloads)'
        in: query
        name: payload
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTracesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or malformed id
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Failed to load session
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Span traces not supported by this backend
      summary: Get a session's trace/span projection
      tags:
      - sessions
...
  /v1/traces/{trace_id}:
    get:
      description: 'Returns one user-visible turn: its spans nested by parent_span_id
        and its dataflow links (links touching other traces included).'
      operationId: getTrace
      parameters:
      - description: Trace id
        in: path
        name: trace_id
        required: true
        schema:
          type: string
      - description: 'Span payload mode: full (default) or preview (strings truncated;
          fetch the span endpoint for full payloads)'
        in: query
        name: payload
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
      summary: Get one trace with spans and links
      tags:
      - traces
...